Use `OwnedFd`/`BorrowedFd` instead of `RawFd`
`Device::new_from_fd()` is unnecessary, since `OwnedFd` and `BorrowedFd`
both implement `AsFd`, so `Device::new()` can be used instead.

This also removes the `'static` bound, which I think is unnecessary. It
should be possible to use a `BorrowedFd`/`&File`/etc. as long as the
`Device` doesn't outlive it.

This raises the minimum Rust version to 1.63.0, assuming there's no
reason to support old versions here. Otherwise, `io-lifetimes` could be
used.